+++ src/features/expressions/shared-tests/functions/displayValue/type-TimePicker.json
@@ -1,0 +1,34 @@
+{
+  "name": "Display value of Timepicker component",
+  "expression": [
+    "displayValue",
+    "tid"
+  ],
+  "context": {
+    "component": "tid",
+    "currentLayout": "Page"
+  },
+  "expects": "03:04",
+  "layouts": {
+    "Page": {
+      "$schema": "https://altinncdn.no/schemas/json/layout/layout.schema.v1.json",
+      "data": {
+        "layout": [
+          {
+            "id": "tid",
+            "type": "Timepicker",
+            "dataModelBindings": {
+              "simpleBinding": "Skjema.Tid"
+            },
+            "format": "HH:mm"
+          }
+        ]
+      }
+    }
+  },
+  "dataModel": {
+    "Skjema": {
+      "Tid": "03:04"
+    }
+  }
+}
